Remove long(id) usage

+++ omero_webtest/views.py
@@ -625,7 +625,7 @@ def image_dimensions(request, image_id, conn=None, **kwargs):
 def image_rois(request, image_id, conn=None, **kwargs):
     """ Simply shows a page of ROI thumbnails for the specified image """
     roi_service = conn.getRoiService()
-    result = roi_service.findByImage(long(image_id), None, conn.SERVICE_OPTS)
+    result = roi_service.findByImage(int(image_id), None, conn.SERVICE_OPTS)
     roi_ids = [r.getId().getValue() for r in result.rois]
     return render(request, 'webtest/demo_viewers/image_rois.html',
                   {'roiIds': roi_ids})
